Hannah - wrote loops to test accuracy of harmonized and unharmonized VGG16 on all images in validation, wrote functions for calculating accuracy in batches and finding variation
Coco - Found and loaded a new dataset for stylized images on imagenet, wrote a function to create a DatasetDict from randomly sampling images from the stylized folder

Stephanie - Worked on creating the manipulated dataset by altering each image, removing high frequency. We created smaller batches to work with so that computation is faster. I created a reusable function to be called in various parts of our project.